refactor: move digest scheduling into whatsapp platform
